feat: major expansion with demo showcase, PTY improvements, and documentation
Comprehensive update adding a full demo showcase application, significant
PTY and terminal improvements, and expanded documentation.

New Demo Showcase Application (crates/ftui-demo-showcase):
- Complete interactive demo with multiple screens:
  - Dashboard with real-time metrics
  - Code explorer with syntax highlighting
  - Data visualization charts
  - File browser with directory navigation
  - Forms and input demonstration
  - Layout laboratory for testing layouts
  - Advanced features showcase
- Standalone binary for demonstrating all frankentui capabilities

PTY and Terminal Improvements (crates/ftui-pty):
- Add +164 lines of PTY handling improvements
- Better process management and signal handling
- Improved output capture and buffering

Core Infrastructure (crates/ftui-core):
- Enhance caps_probe.rs with terminal capability detection (+26 lines)
- Update terminal_session.rs with session management improvements
- Refactor input_parser.rs for cleaner code

Widget and Extras Updates:
- charts.rs: Improved chart rendering (+34 lines)
- clipboard.rs: Additional clipboard handling
- markdown.rs: Enhanced markdown parsing (+41 lines)
- syntax.rs: Syntax highlighting refinements
- search.rs: Text search improvements (+88 lines)
- Various widget updates (borders, input, list, json_view)

Runtime Improvements:
- log_sink.rs: Better log handling (+23 lines)
- program.rs: Program lifecycle improvements (+24 lines)
- subscription.rs: Enhanced event subscriptions (+22 lines)
- terminal_writer.rs: Major terminal output improvements (+114 lines)

Documentation:
- Significant README.md expansion (+370 lines)
- Updated FIXES_SUMMARY.md and SESSION_TODO.md

+++ crates/ftui-widgets/src/json_view.rs
@@ -329,7 +329,7 @@ impl Widget for JsonView {
         let max_x = area.right();
 
         for (row_idx, tokens) in lines.iter().enumerate() {
-            if row_idx as u16 >= area.height {
+            if row_idx >= area.height as usize {
                 break;
             }
